Get score tracking and Question component functionality working
